requirements.txt
djangorestframework~=3.12
pillow~=9.4
django~=4.1
django-cors-headers~=3.14
python-dotenv~=0.21
